+++ prnt_shuint.c
@@ -11,7 +11,7 @@ int prnt_shuint(va_list the_list, char *buffp, unsigned int buffndx)
 {
     unsigned short int inp_in, inp_tmp, iden, dive;
 
-    inp_in = va_arg(the_list, unsigned short int);
+    inp_in = va_arg(the_list, unsigned int);
     inp_tmp = inp_in;
     dive = 1;
     while (inp_tmp > 9)
